+++ src/Utils/ServerAwareTrait.php
@@ -1,0 +1,38 @@
+<?php
+namespace League\OAuth2\Server\Utils;
+
+use League\OAuth2\Server\AbstractServer;
+
+trait ServerAwareTrait
+{
+    /**
+     * Server
+     *
+     * @var \League\OAuth2\Server\AbstractServer $server
+     */
+    protected $server;
+
+    /**
+     * Set the server
+     *
+     * @param \League\OAuth2\Server\AbstractServer $server
+     *
+     * @return self
+     */
+    public function setServer(AbstractServer $server)
+    {
+        $this->server = $server;
+
+        return $this;
+    }
+
+    /**
+     * Return the server
+     *
+     * @return \League\OAuth2\Server\AbstractServer
+     */
+    protected function getServer()
+    {
+        return $this->server;
+    }
+}
